[FIX] remove filter[has_delivery_order] SO & update role

+++ app/Http/Controllers/Api/RoleController.php
@@ -23,7 +23,6 @@ class RoleController extends Controller
             ->with('permissions')
             ->allowedFilters(['name'])
             ->allowedSorts(['id', 'name', 'created_at'])
-            ->orderBy('id', 'DESC')
             ->paginate();
 
         return RoleResource::collection($roles);
@@ -37,6 +36,7 @@ class RoleController extends Controller
      */
     public function store(RoleStoreRequest $request)
     {
+        abort_if(!auth()->user()->tokenCan('role_create'), 403);
         $role = new Role();
         $role->name = $request->name;
         $role->guard_name = 'web';
@@ -66,6 +66,8 @@ class RoleController extends Controller
      */
     public function update(Role $role, RoleUpdateRequest $request)
     {
+        abort_if(!auth()->user()->tokenCan('role_edit'), 403);
+        if ($role->id == 1) return response()->json(['message' => 'Role admin can not updated!']);
         $role->name = $request->input('name');
         $role->save();
         $role->syncPermissions($request->permission_ids ?? []);
@@ -81,6 +83,7 @@ class RoleController extends Controller
     public function destroy(Role $role)
     {
         abort_if(!auth()->user()->tokenCan('role_delete'), 403);
+        if ($role->id == 1) return response()->json(['message' => 'Role admin can not deleted!']);
         $role->delete();
         return $this->deletedResponse();
     }
